fix(omo-ai): address final review blockers
