+++ packages/omo-ai/src/install/postinstall.ts
@@ -1,5 +1,5 @@
-import { getInstallMode } from "./index.ts";
+import { OMO_AI_PACKAGE_VERSION } from "../senpi-compat/index.ts";
 
 if (import.meta.main) {
-  console.log(`omo-ai postinstall ${getInstallMode()}`);
+  console.log(`omo-ai postinstall ${OMO_AI_PACKAGE_VERSION}`);
 }
