+++ packages/omo-ai/test/fixtures/final-review-evidence.md
@@ -1,0 +1,17 @@
+# omo-ai final review evidence
+
+Evidence root: `.omo/evidence/20260701-omo-ai-senpi-package/final-review/split-tests/`.
+
+This receipt records the final review pass after splitting the oversized installer CLI test into behavior-cluster files and fixing the package export blocker. Current evidence supersedes older REJECT and blocked artifacts from earlier review attempts; those artifacts remain historical only.
+
+Coverage recorded:
+
+- `cd packages/omo-ai && bun run test`
+- `/opt/homebrew/bin/tsc --noEmit -p packages/omo-ai/tsconfig.json`
+- package import probes for `omo-ai/install` and `omo-ai/doctor`, proving the public subpaths expose the real runtime functions
+- wrong-shape JSON boundary tests for `settings.json` and `hooks-state.json`, proving repair fails before mutation
+- pure LOC scan for `packages/omo-ai/test/*.test.ts`, proving every test file is at or below 250 nonblank, noncomment lines
+- root `typecheck:packages` script proof that the new workspace is included in the repo-level package typecheck chain
+- cleanup receipt confirming no task-created temp directories remained
+
+The full root `bun run typecheck:packages` command is recorded separately and currently fails before reaching `omo-ai` on pre-existing `packages/rules-engine` Node type-resolution errors. The targeted `omo-ai` typecheck is the package-specific pass for this change.
